requirements.txt
# --- Core ---
PyYAML==6.0.2
requests==2.32.3
pydub==0.25.1
Jinja2==3.1.4
Flask==3.0.3

# --- Browser automation ---
selenium==4.23.1
webdriver-manager==4.0.2

# --- Video/audio processing ---
opencv-python==4.10.0.84
numpy==1.26.4
scipy==1.11.4
librosa==0.10.1
tqdm==4.66.4
numba==0.59.1

# --- Face alignment (Wav2Lip deps) ---
face-alignment==1.3.5

# --- PyTorch CPU build (change index_url for GPU) ---
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu
-f https://download.pytorch.org/whl/cpu/torch_stable.html

# --- Optional: improve audio features ---
soundfile==0.12.1
